src/main/java/com/jmms/api/domain/common/ResponseInformation.java
package com.jmms.api.domain.common;


public class ResponseInformation<T> {

    private Application application;
    private Request request;
    private T response;

    public Application getApplication() {
        return application;
    }

    public void setApplication(Application application) {
        this.application = application;
    }

    public Request getRequest() {
        return request;
    }

    public void setRequest(Request request) {
        this.request = request;
    }

    public T getResponse() {
        return response;
    }

    public void setResponse(T response) {
        this.response = response;
    }

}
